got mdx to render properly
