+++ gatsby-node.js
@@ -40,7 +40,7 @@ exports.createPages = async ({ graphql, actions }) => {
   result.data.allMdx.edges.forEach(({ node }) => {
     createPage({
       path: node.fields.slug,
-      component: path.resolve(`./src/components/react/layout.js`),
+      component: path.resolve(`./src/templates/react_template.js`),
       context: {
         slug: node.fields.slug
       }
