+++ src/components/layout.js
@@ -5,22 +5,21 @@
  * See: https://www.gatsbyjs.org/docs/use-static-query/
  */
 
-import React from "react";
-import PropTypes from "prop-types";
-import { useStaticQuery, graphql } from "gatsby";
+import React from 'react';
+import PropTypes from 'prop-types';
 
-import Header from "./header";
-import "./styles/layout.css";
+import Header from './header';
+import './styles/layout.css';
 
-import { makeStyles } from "@material-ui/core/styles";
+import { makeStyles } from '@material-ui/core/styles';
 
-import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
+import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
+import Container from '@material-ui/core/Container';
 
-const primary = "#004d84";
-const secondary = "#05a68e";
+const primary = '#004d84';
+const secondary = '#05a68e';
 
-import Paper from "@material-ui/core/Paper";
+import Paper from '@material-ui/core/Paper';
 
 const theme = createMuiTheme({
   palette: {
@@ -29,8 +28,8 @@ const theme = createMuiTheme({
       main: secondary
     },
     text: {
-      primary: "#fff",
-      secondary: "#000000"
+      primary: '#fff',
+      secondary: '#000000'
     }
   },
   spacing: 2
@@ -43,6 +42,7 @@ const useStyles = makeStyles(theme => ({
 
 export default function Layout({ children }) {
   const classes = useStyles(theme);
+
   return (
     <>
       <ThemeProvider theme={theme}>
